refactor: 공통 ViewModelFactory 구현
