+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/KerdyApplication.kt
@@ -1,0 +1,16 @@
+package com.emmsale.presentation
+
+import android.app.Application
+import android.content.Context
+
+class KerdyApplication : Application() {
+
+    init {
+        instance = this
+    }
+
+    companion object {
+        private lateinit var instance: KerdyApplication
+        val context: Context get() = instance.applicationContext
+    }
+}
