+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/common/keys/Keys.kt
@@ -1,3 +1,3 @@
-package com.emmsale.presentation.utils.keys
+package com.emmsale.presentation.common.keys
 
 internal const val KERDY_PREF_KEY = "kerdy_pref_key"
