+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/common/livedata/ListLiveData.kt
@@ -1,29 +1,19 @@
-package com.emmsale.presentation.utils.livedata
+package com.emmsale.presentation.common.livedata
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.map
 
-class DistinctListLiveData<T> : MutableLiveData<MutableList<T>>() {
-    private var items = mutableListOf<T>()
+class ListLiveData<T> : MutableLiveData<MutableList<T>>() {
+    private val items = mutableListOf<T>()
 
     fun add(item: T) {
         items.add(item)
-        val distinctItems = items.distinct().toMutableList()
-        items.clear()
-        items.addAll(distinctItems)
-
-        postValue(distinctItems)
         updateState()
     }
 
     fun addAll(newItems: List<T>) {
         items.addAll(newItems)
-        val distinctItems = items.distinct().toMutableList()
-        items.clear()
-        items.addAll(distinctItems)
-
-        postValue(distinctItems)
         updateState()
     }
 
@@ -43,7 +33,7 @@ class DistinctListLiveData<T> : MutableLiveData<MutableList<T>>() {
     }
 
     private fun updateState() {
-        value = items
+        postValue(items)
     }
 
     fun asLiveData(): LiveData<List<T>> = this.map { it }
