+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/onboarding/OnboardingNameFragment.kt
@@ -6,9 +6,10 @@ import androidx.fragment.app.activityViewModels
 import com.emmsale.R
 import com.emmsale.databinding.FragmentOnboardingNameBinding
 import com.emmsale.presentation.base.fragment.BaseFragment
+import com.emmsale.presentation.common.viewModelFactory
 
 class OnboardingNameFragment : BaseFragment<FragmentOnboardingNameBinding>() {
-    override val viewModel: OnboardingViewModel by activityViewModels { OnboardingViewModelFactory() }
+    override val viewModel: OnboardingViewModel by activityViewModels { viewModelFactory }
     override val layoutResId: Int = R.layout.fragment_onboarding_name
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
